Restore optimize structure
