+++ src/optimize/index.ts
@@ -1,7 +1,7 @@
-import { walk, asyncWalk } from 'estree-walker';
+import { walk } from 'estree-walker';
 import type { Ast, TemplateNode } from '../compiler/interfaces';
 import { NodeVisitor, Optimizer, VisitorFn } from '../@types/optimizer';
-import { StyleTransformResult, transformStyle } from './style.js';
+import optimizeStyles from './styles.js';
 
 interface VisitorCollection {
   enter: Map<string, VisitorFn[]>;
@@ -9,13 +9,12 @@ interface VisitorCollection {
 }
 
 function addVisitor(visitor: NodeVisitor, collection: VisitorCollection, nodeName: string, event: 'enter' | 'leave') {
-  if (event in visitor) {
-    if (collection[event].has(nodeName)) {
-      collection[event].get(nodeName)!.push(visitor[event]!);
-    }
+  if (typeof visitor[event] !== 'function') return;
+  if (!collection[event]) collection[event] = new Map<string, VisitorFn[]>();
 
-    collection.enter.set(nodeName, [visitor[event]!]);
-  }
+  const visitors = collection[event].get(nodeName) || [];
+  visitors.push(visitor[event] as any);
+  collection[event].set(nodeName, visitors);
 }
 
 function collectVisitors(optimizer: Optimizer, htmlVisitors: VisitorCollection, cssVisitors: VisitorCollection, finalizers: Array<() => Promise<void>>) {
@@ -70,97 +69,15 @@ interface OptimizeOptions {
 }
 
 export async function optimize(ast: Ast, opts: OptimizeOptions) {
-  try {
-    const classNames: Set<string> = new Set();
-    const allCssModules: Map<string, string> = new Map();
-    let hasStyles = false;
+  const htmlVisitors = createVisitorCollection();
+  const cssVisitors = createVisitorCollection();
+  const finalizers: Array<() => Promise<void>> = [];
 
-    // 1. collect all CSS classes in HTML. Must be completely done before styles are parsed.
-    walk(ast.html, {
-      enter(node) {
-        if (node.type !== 'Element') return;
-        for (let attr of node.attributes) {
-          if (attr.name !== 'class') continue;
-          for (let value of attr.value) {
-            if (value.type !== 'Text') continue;
-            (value.data as string)
-              .split(' ')
-              .map((c) => c.trim())
-              .forEach((c) => classNames.add(c));
-          }
-        }
-      },
-    });
+  collectVisitors(optimizeStyles(opts), htmlVisitors, cssVisitors, finalizers);
 
-    // 2. parse styles
-    await asyncWalk(ast.css, {
-      async enter(node: TemplateNode) {
-        if (node.type !== 'Style') return;
-        // TODO: check for top-level <style> tag; ignore within component
+  walkAstWithVisitors(ast.css, cssVisitors);
+  walkAstWithVisitors(ast.html, htmlVisitors);
 
-        hasStyles = true; // some ASTs don’t have any styles; skip step 3 if so
-        const code = node.content.styles;
-        const typeAttr = node.attributes && node.attributes.find(({ name }: { name: string }) => name === 'type');
-
-        // 1. transform styles
-        const result = await transformStyle(code, {
-          type: (typeAttr.value[0] && typeAttr.value[0].raw) || undefined,
-          classNames,
-          filename: opts.filename,
-          fileID: opts.fileID,
-        });
-
-        // 2. save cssModule names for updating HTML below
-        for (const [k, v] of result.cssModules) {
-          allCssModules.set(k, v);
-        }
-
-        // 3. Update CSS AST (including <style type="">)
-        node.content.styles = result.css;
-        if (result.type !== 'text/css') {
-          for (let i = 0; i < node.attributes.length; i++) {
-            if (node.attributes[i].name !== 'type') continue;
-            for (let j = 0; j < node.attributes[i].value.length; j++) {
-              node.attributes[i].value[j].raw = 'text/css';
-              node.attributes[i].value[j].data = 'text/css';
-            }
-          }
-        }
-
-        this.replace(node);
-      },
-    });
-
-    // 3. Update HTML with new local classnames and inject styles
-    if (hasStyles) {
-      if (allCssModules.size) {
-        walk(ast.html, {
-          enter(node) {
-            if (node.type === 'Element') {
-              for (let i = 0; i < node.attributes.length; i++) {
-                if (node.attributes[i].name !== 'class') continue;
-                for (let j = 0; j < node.attributes[i].value.length; j++) {
-                  if (node.attributes[i].value[j].type !== 'Text') continue;
-                  const elementClassNames = (node.attributes[i].value[j].raw as string)
-                    .split(' ')
-                    .map((c) => {
-                      let className = c.trim();
-                      return allCssModules.get(className) || className;
-                    })
-                    .join(' ');
-                  node.attributes[i].value[j].raw = elementClassNames;
-                  node.attributes[i].value[j].data = elementClassNames;
-                }
-              }
-              this.replace(node);
-            }
-          },
-        });
-      }
-    }
-
-    // Done! The classNames are updated in the HTML, but we’ll inject the styles later
-  } catch (err) {
-    console.error(err);
-  }
+  // Run all of the finalizer functions in parallel because why not.
+  await Promise.all(finalizers.map((fn) => fn()));
 }
